backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "sqlmodel>=0.0.24",
    "psycopg[binary]>=3.1",
    "uvicorn[standard]>=0.35.0",
    "alembic>=1.13.0",
    "pgvector>=0.4.1",
    "google-auth>=2.23.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.1.0",
    "pyjwt>=2.8.0",
    "cryptography>=41.0.0",
    "httpx>=0.24.0",
]
